Revert "changed convolution to use c2c, not r2c and c2r but it dooesnt' work"
This reverts commit 97aac071330e893bd4c67c52fa2bcccc6fd59f4b.

+++ CudaProjectImageProcessing/ConvolutionOld.cuh
@@ -5,7 +5,7 @@ class ConvolutionCalculator
 public:
 	static void convolution(
 		const uint8_t* image_in,
-		const double* kernel,
+		const float* kernel,
 		const int image_width,
 		const int image_height,
 		const int kernel_width,
